+++ packages/atomic-react/src/components/search/components.ts
@@ -5,12 +5,16 @@ import {
   AtomicFacet as LitAtomicFacet,
   AtomicHtml as LitAtomicHtml,
   AtomicIcon as LitAtomicIcon,
+  AtomicInsightGenerateAnswerButton as LitAtomicInsightGenerateAnswerButton,
   AtomicInsightInterface as LitAtomicInsightInterface,
   AtomicLayoutSection as LitAtomicLayoutSection,
+  AtomicLoadMoreResults as LitAtomicLoadMoreResults,
   AtomicPager as LitAtomicPager,
   AtomicQuerySummary as LitAtomicQuerySummary,
+  AtomicRecsInterface as LitAtomicRecsInterface,
   AtomicResult as LitAtomicResult,
   AtomicResultChildrenTemplate as LitAtomicResultChildrenTemplate,
+  AtomicResultList as LitAtomicResultList,
   AtomicResultSectionActions as LitAtomicResultSectionActions,
   AtomicResultSectionBadges as LitAtomicResultSectionBadges,
   AtomicResultSectionBottomMetadata as LitAtomicResultSectionBottomMetadata,
@@ -21,11 +25,13 @@ import {
   AtomicResultSectionTitleMetadata as LitAtomicResultSectionTitleMetadata,
   AtomicResultSectionVisual as LitAtomicResultSectionVisual,
   AtomicResultsPerPage as LitAtomicResultsPerPage,
+  AtomicResultText as LitAtomicResultText,
   AtomicSearchBoxInstantResults as LitAtomicSearchBoxInstantResults,
   AtomicSearchBoxQuerySuggestions as LitAtomicSearchBoxQuerySuggestions,
   AtomicSearchBoxRecentQueries as LitAtomicSearchBoxRecentQueries,
   AtomicSearchInterface as LitAtomicSearchInterface,
   AtomicSearchLayout as LitAtomicSearchLayout,
+  AtomicText as LitAtomicText,
 } from '@coveo/atomic/components';
 import {createComponent} from '@lit/react';
 import React from 'react';
@@ -66,6 +72,12 @@ export const AtomicIcon = createComponent({
   elementClass: LitAtomicIcon,
 });
 
+export const AtomicInsightGenerateAnswerButton = createComponent({
+  tagName: 'atomic-insight-generate-answer-button',
+  react: React,
+  elementClass: LitAtomicInsightGenerateAnswerButton,
+});
+
 export const AtomicInsightInterface = createComponent({
   tagName: 'atomic-insight-interface',
   react: React,
@@ -76,6 +88,12 @@ export const AtomicLayoutSection = createComponent({
   tagName: 'atomic-layout-section',
   react: React,
   elementClass: LitAtomicLayoutSection,
+});
+
+export const AtomicLoadMoreResults = createComponent({
+  tagName: 'atomic-load-more-results',
+  react: React,
+  elementClass: LitAtomicLoadMoreResults,
 });
 
 export const AtomicPager = createComponent({
@@ -90,6 +108,12 @@ export const AtomicQuerySummary = createComponent({
   elementClass: LitAtomicQuerySummary,
 });
 
+export const AtomicRecsInterface = createComponent({
+  tagName: 'atomic-recs-interface',
+  react: React,
+  elementClass: LitAtomicRecsInterface,
+});
+
 export const AtomicResult = createComponent({
   tagName: 'atomic-result',
   react: React,
@@ -100,6 +124,12 @@ export const AtomicResultChildrenTemplate = createComponent({
   tagName: 'atomic-result-children-template',
   react: React,
   elementClass: LitAtomicResultChildrenTemplate,
+});
+
+export const AtomicResultList = createComponent({
+  tagName: 'atomic-result-list',
+  react: React,
+  elementClass: LitAtomicResultList,
 });
 
 export const AtomicResultSectionActions = createComponent({
@@ -162,6 +192,12 @@ export const AtomicResultsPerPage = createComponent({
   elementClass: LitAtomicResultsPerPage,
 });
 
+export const AtomicResultText = createComponent({
+  tagName: 'atomic-result-text',
+  react: React,
+  elementClass: LitAtomicResultText,
+});
+
 export const AtomicSearchBoxInstantResults = createComponent({
   tagName: 'atomic-search-box-instant-results',
   react: React,
@@ -190,4 +226,10 @@ export const AtomicSearchLayout = createComponent({
   tagName: 'atomic-search-layout',
   react: React,
   elementClass: LitAtomicSearchLayout,
+});
+
+export const AtomicText = createComponent({
+  tagName: 'atomic-text',
+  react: React,
+  elementClass: LitAtomicText,
 });