+++ packages/atomic/src/components/commerce/atomic-commerce-did-you-mean/e2e/atomic-commerce-did-you-mean.e2e.ts
@@ -6,11 +6,6 @@ test.describe('atomic-commerce-did-you-mean', () => {
     await commerceDidYouMean.hydrated.waitFor();
   });
 
-  test('should be accessible', async ({makeAxeBuilder}) => {
-    const accessibilityResults = await makeAxeBuilder().analyze();
-    expect(accessibilityResults.violations).toEqual([]);
-  });
-
   test('should display the auto correction message', async ({
     commerceDidYouMean,
   }) => {