+++ packages/atomic/src/components/commerce/atomic-commerce-query-error/atomic-commerce-query-error.new.stories.tsx
@@ -1,12 +1,12 @@
 import type {Meta, StoryObj as Story} from '@storybook/web-components-vite';
 import {getStorybookHelpers} from '@wc-toolkit/storybook-helpers';
-import {HttpResponse, http} from 'msw';
+import {MockCommerceApi} from '@/storybook-utils/api/commerce/mock';
 import {wrapInCommerceInterface} from '@/storybook-utils/commerce/commerce-interface-wrapper';
 import {parameters} from '@/storybook-utils/common/common-meta-parameters';
 
-const {decorator, play} = wrapInCommerceInterface({
-  engineConfig: {organizationId: 'invalid-organization-id'},
-});
+const mockCommerceApi = new MockCommerceApi();
+
+const {decorator, play} = wrapInCommerceInterface();
 const {events, args, argTypes, template} = getStorybookHelpers(
   'atomic-commerce-query-error',
   {excludeCategories: ['methods']}
@@ -23,36 +23,36 @@ const meta: Meta = {
     actions: {
       handles: events,
     },
+    msw: {handlers: [...mockCommerceApi.handlers]},
   },
   args,
   argTypes,
-
+  beforeEach: async () => {
+    mockCommerceApi.searchEndpoint.clear();
+  },
   play,
 };
 
 export default meta;
 
-export const Default: Story = {};
+export const Default: Story = {
+  beforeEach: async () => {
+    mockCommerceApi.searchEndpoint.mockErrorOnce();
+  },
+};
 
 export const With418Error: Story = {
   name: 'With 418 error',
-  parameters: {
-    msw: {
-      handlers: [
-        http.post('**/v2/search', () => {
-          return HttpResponse.json(
-            {
-              ok: false,
-              status: 418,
-              message: 'Something very weird just happened',
-              statusCode: 418,
-              type: 'ClientError',
-            },
-            {status: 418}
-          );
-        }),
-      ],
-    },
+  beforeEach: async () => {
+    mockCommerceApi.searchEndpoint.mockOnce(
+      () => ({
+        ok: false,
+        status: 418,
+        statusCode: 418,
+        message: 'Something very weird just happened',
+        type: 'ClientError',
+      }),
+      {status: 418}
+    );
   },
-  play,
 };