+++ packages/atomic/src/components/commerce/atomic-commerce-search-box-query-suggestions/e2e/fixture.ts
@@ -1,13 +1,11 @@
 import {test as base} from '@playwright/test';
-import {type AxeFixture, makeAxeBuilder} from '@/playwright-utils/base-fixture';
 import {AtomicCommerceSearchBoxQuerySuggestionsPageObject} from './page-object';
 
 type Fixtures = {
   commerceSearchBoxQuerySuggestions: AtomicCommerceSearchBoxQuerySuggestionsPageObject;
 };
 
-export const test = base.extend<Fixtures & AxeFixture>({
-  makeAxeBuilder,
+export const test = base.extend<Fixtures>({
   commerceSearchBoxQuerySuggestions: async ({page}, use) => {
     await use(new AtomicCommerceSearchBoxQuerySuggestionsPageObject(page));
   },