+++ packages/headless/src/controllers/insight/recent-queries-list/headless-insight-recent-queries-list.test.ts
@@ -99,6 +99,9 @@ describe('insight recent queries list', () => {
       expect(mockCoreController.executeRecentQuery).toHaveBeenCalledWith(index);
       expect(executeSearch).toHaveBeenCalledWith({
         legacy: logRecentQueryClick(),
+        next: {
+          actionCause: 'recentQueriesClick',
+        },
       });
     });
   });